extend yurt-iot-dock to sync it with the configuration in PlatformAdmin
