+++ pkg/controller/platformadmin/util.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package platformadmin
 
 import (
 	"fmt"
@@ -28,71 +28,40 @@ import (
 
 	iotv1alpha2 "github.com/openyurtio/openyurt/pkg/apis/iot/v1alpha2"
 	"github.com/openyurtio/openyurt/pkg/controller/platformadmin/config"
+	utils "github.com/openyurtio/openyurt/pkg/controller/platformadmin/utils"
 )
 
-// NewPlatformAdminCondition creates a new PlatformAdmin condition.
-func NewPlatformAdminCondition(condType iotv1alpha2.PlatformAdminConditionType, status corev1.ConditionStatus, reason, message string) *iotv1alpha2.PlatformAdminCondition {
-	return &iotv1alpha2.PlatformAdminCondition{
-		Type:               condType,
-		Status:             status,
-		LastTransitionTime: metav1.Now(),
-		Reason:             reason,
-		Message:            message,
-	}
-}
-
-// GetPlatformAdminCondition returns the condition with the provided type.
-func GetPlatformAdminCondition(status iotv1alpha2.PlatformAdminStatus, condType iotv1alpha2.PlatformAdminConditionType) *iotv1alpha2.PlatformAdminCondition {
-	for i := range status.Conditions {
-		c := status.Conditions[i]
-		if c.Type == condType {
-			return &c
-		}
-	}
-	return nil
-}
-
-// SetPlatformAdminCondition updates the PlatformAdmin to include the provided condition. If the condition that
-// we are about to add already exists and has the same status, reason and message then we are not going to update.
-func SetPlatformAdminCondition(status *iotv1alpha2.PlatformAdminStatus, condition *iotv1alpha2.PlatformAdminCondition) {
-	currentCond := GetPlatformAdminCondition(*status, condition.Type)
-	if currentCond != nil && currentCond.Status == condition.Status && currentCond.Reason == condition.Reason {
-		return
-	}
-
-	if currentCond != nil && currentCond.Status == condition.Status {
-		condition.LastTransitionTime = currentCond.LastTransitionTime
-	}
-	newConditions := filterOutCondition(status.Conditions, condition.Type)
-	status.Conditions = append(newConditions, *condition)
-}
-
-func filterOutCondition(conditions []iotv1alpha2.PlatformAdminCondition, condType iotv1alpha2.PlatformAdminConditionType) []iotv1alpha2.PlatformAdminCondition {
-	var newConditions []iotv1alpha2.PlatformAdminCondition
-	for _, c := range conditions {
-		if c.Type == condType {
-			continue
-		}
-		newConditions = append(newConditions, c)
-	}
-	return newConditions
-}
-
-func NewYurtIoTDockComponent(platformAdmin *iotv1alpha2.PlatformAdmin) (*config.Component, error) {
+// NewYurtIoTDockComponent initialize the configuration of yurt-iot-dock component
+func NewYurtIoTDockComponent(platformAdmin *iotv1alpha2.PlatformAdmin, platformAdminFramework *PlatformAdminFramework) (*config.Component, error) {
 	var yurtIotDockComponent config.Component
 
-	ver, ns, err := defaultVersion(platformAdmin)
+	// If the configuration of the yurt-iot-dock component that customized in the platformAdminFramework
+	for _, cp := range platformAdminFramework.Components {
+		if cp.Name != utils.IotDockName {
+			continue
+		}
+		return cp, nil
+	}
+
+	// Otherwise, the default configuration is used to start
+	ver, ns, err := utils.DefaultVersion(platformAdmin)
 	if err != nil {
 		return nil, err
 	}
 
-	yurtIotDockComponent.Name = IotDockName
+	yurtIotDockComponent.Name = utils.IotDockName
 	yurtIotDockComponent.Deployment = &appsv1.DeploymentSpec{
+		Selector: &metav1.LabelSelector{
+			MatchLabels: map[string]string{
+				"app":           "yurt-iot-dock",
+				"control-plane": "platformadmin-controller",
+			},
+		},
 		Template: corev1.PodTemplateSpec{
 			ObjectMeta: metav1.ObjectMeta{
 				Labels: map[string]string{
 					"app":           "yurt-iot-dock",
-					"control-plane": "edgex-controller-manager",
+					"control-plane": "platformadmin-controller",
 				},
 				Namespace: ns,
 			},
